bdf:  - fixing DRESP2 check on DTABLE
op2_geom:
 - various NX fixes found by bad data length:
   - TSTEPNL, TLOAD1, TLOAD2, MAT10, CTRIAX, EIGB
 - 64 bit support:
   - SEBULK, SPOINT, MAT10, MATT2, MATT9

+++ pyNastran/bdf/bdf.py
@@ -3239,7 +3239,7 @@ class BDF_(BDFMethods, GetCard, AddCards, WriteMeshs, UnXrefMesh):
 
         Examples
         --------
-        >>> out = model.get_xyz_in_coord_array(cid=0)
+        >>> out = model.get_xyz_in_coord_array(cid=0, fdtype='float64', idtype='int32')
         >>> nid_cp_cd, xyz_cid, xyz_cp, icd_transform, icp_transform = out
         """
         icd_transform, icp_transform, xyz_cp, nid_cp_cd = self.get_displacement_index_xyz_cp_cd(
